// packages/cactus-plugin-keychain-memory/src/main/typescript/generated/crpc/models/has_keychain_entry_response_v1_pb_pb.ts
//
//Hyperledger Cactus Plugin - Keychain Memory 
//
//Contains/describes the Hyperledger Cacti Keychain Memory plugin.
//
//The version of the OpenAPI document: 2.1.0
//
//Generated by OpenAPI Generator: https://openapi-generator.tech

// @generated by protoc-gen-es v2.11.0 with parameter "target=ts"
// @generated from file models/has_keychain_entry_response_v1_pb.proto (package org.hyperledger.cacti.plugin.keychain.memory, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import { file_google_protobuf_any } from "@bufbuild/protobuf/wkt";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file models/has_keychain_entry_response_v1_pb.proto.
 */
export const file_models_has_keychain_entry_response_v1_pb: GenFile = /*@__PURE__*/
  fileDesc("Ci5tb2RlbHMvaGFzX2tleWNoYWluX2VudHJ5X3Jlc3BvbnNlX3YxX3BiLnByb3RvEixvcmcuaHlwZXJsZWRnZXIuY2FjdGkucGx1Z2luLmtleWNoYWluLm1lbW9yeSJbChxIYXNLZXljaGFpbkVudHJ5UmVzcG9uc2VWMVBCEg0KA2tleRjfvAYgASgJEhUKCWNoZWNrZWRBdBi6lKa+ASABKAkSFQoJaXNQcmVzZW50GND/nKwBIAEoCGIGcHJvdG8z", [file_google_protobuf_any]);

/**
 * @generated from message org.hyperledger.cacti.plugin.keychain.memory.HasKeychainEntryResponseV1PB
 */
export type HasKeychainEntryResponseV1PB = Message<"org.hyperledger.cacti.plugin.keychain.memory.HasKeychainEntryResponseV1PB"> & {
  /**
   * The key that was used to check the presence of the value in the entry store.
   *
   * @generated from field: string key = 106079;
   */
  key: string;

  /**
   * Date and time encoded as JSON when the presence check was performed by the plugin backend.
   *
   * @generated from field: string checkedAt = 399084090;
   */
  checkedAt: string;

  /**
   * The boolean true or false indicating the presence or absence of an entry under 'key'.
   *
   * @generated from field: bool isPresent = 361185232;
   */
  isPresent: boolean;
};

/**
 * Describes the message org.hyperledger.cacti.plugin.keychain.memory.HasKeychainEntryResponseV1PB.
 * Use `create(HasKeychainEntryResponseV1PBSchema)` to create a new message.
 */
export const HasKeychainEntryResponseV1PBSchema: GenMessage<HasKeychainEntryResponseV1PB> = /*@__PURE__*/
  messageDesc(file_models_has_keychain_entry_response_v1_pb, 0);