// packages/cactus-plugin-keychain-memory/src/main/typescript/generated/crpc/models/set_keychain_entry_response_v1_pb_pb.ts
//
//Hyperledger Cactus Plugin - Keychain Memory 
//
//Contains/describes the Hyperledger Cacti Keychain Memory plugin.
//
//The version of the OpenAPI document: 2.1.0
//
//Generated by OpenAPI Generator: https://openapi-generator.tech

// @generated by protoc-gen-es v2.11.0 with parameter "target=ts"
// @generated from file models/set_keychain_entry_response_v1_pb.proto (package org.hyperledger.cacti.plugin.keychain.memory, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import { file_google_protobuf_any } from "@bufbuild/protobuf/wkt";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file models/set_keychain_entry_response_v1_pb.proto.
 */
export const file_models_set_keychain_entry_response_v1_pb: GenFile = /*@__PURE__*/
  fileDesc("Ci5tb2RlbHMvc2V0X2tleWNoYWluX2VudHJ5X3Jlc3BvbnNlX3YxX3BiLnByb3RvEixvcmcuaHlwZXJsZWRnZXIuY2FjdGkucGx1Z2luLmtleWNoYWluLm1lbW9yeSItChxTZXRLZXljaGFpbkVudHJ5UmVzcG9uc2VWMVBCEg0KA2tleRjfvAYgASgJYgZwcm90bzM", [file_google_protobuf_any]);

/**
 * @generated from message org.hyperledger.cacti.plugin.keychain.memory.SetKeychainEntryResponseV1PB
 */
export type SetKeychainEntryResponseV1PB = Message<"org.hyperledger.cacti.plugin.keychain.memory.SetKeychainEntryResponseV1PB"> & {
  /**
   * The key that was used to set the value on the keychain.
   *
   * @generated from field: string key = 106079;
   */
  key: string;
};

/**
 * Describes the message org.hyperledger.cacti.plugin.keychain.memory.SetKeychainEntryResponseV1PB.
 * Use `create(SetKeychainEntryResponseV1PBSchema)` to create a new message.
 */
export const SetKeychainEntryResponseV1PBSchema: GenMessage<SetKeychainEntryResponseV1PB> = /*@__PURE__*/
  messageDesc(file_models_set_keychain_entry_response_v1_pb, 0);